Updated examples and snippets for rtu v2
