+++ ScanbotBarcodeSDKExample/BarcodeSDK.NET.Droid.Example/Resources/layout/activity_main.xml
@@ -138,6 +138,22 @@
 				android:layout_height="1dp"
 				android:background="#50000000" />
 
+            <TextView
+				android:id="@+id/rtu_ui_v2_single_scanning"
+				android:layout_width="match_parent"
+				android:layout_height="wrap_content"
+				android:gravity="center_vertical"
+				android:paddingHorizontal="16dp"
+				android:paddingVertical="20dp"
+				style="@style/Scanbot.Demo.MaterialTwoButtonTeaser.Button"
+				android:text="RTU UI - single scanning"
+				android:textColor="@color/primary_material_light" />
+
+			<FrameLayout
+				android:layout_width="match_parent"
+				android:layout_height="1dp"
+				android:background="#50000000" />
+
 			<TextView
 				android:id="@+id/rtu_ui_v2_aroverlay"
 				android:layout_width="match_parent"
@@ -146,7 +162,7 @@
 				android:paddingHorizontal="16dp"
 				android:paddingVertical="20dp"
 				style="@style/Scanbot.Demo.MaterialTwoButtonTeaser.Button"
-				android:text="Barcode Scanner ArOverlay"
+				android:text="RTU UI - with selection overlay"
 				android:textColor="@color/primary_material_light" />
 
 			<FrameLayout
@@ -178,7 +194,7 @@
 				android:paddingHorizontal="16dp"
 				android:paddingVertical="20dp"
 				style="@style/Scanbot.Demo.MaterialTwoButtonTeaser.Button"
-				android:text="Barcode Scanner MultipleScanning Preview"
+				android:text="RTU UI - multiple unique barcode scanning"
 				android:textColor="@color/primary_material_light" />
 
 			<FrameLayout
@@ -194,7 +210,7 @@
 				android:paddingHorizontal="16dp"
 				android:paddingVertical="20dp"
 				style="@style/Scanbot.Demo.MaterialTwoButtonTeaser.Button"
-				android:text="Barcode Scanner MultipleScanning"
+				android:text="RTU UI - find and pick barcode scanning"
 				android:textColor="@color/primary_material_light" />
 
 			<FrameLayout
@@ -202,22 +218,6 @@
 				android:layout_height="1dp"
 				android:background="#50000000" />
 			
-			<TextView
-				android:id="@+id/rtu_ui_v2_single_scanning"
-				android:layout_width="match_parent"
-				android:layout_height="wrap_content"
-				android:gravity="center_vertical"
-				android:paddingHorizontal="16dp"
-				android:paddingVertical="20dp"
-				style="@style/Scanbot.Demo.MaterialTwoButtonTeaser.Button"
-				android:text="Barcode Scanner SingleScanning"
-				android:textColor="@color/primary_material_light" />
-
-			<FrameLayout
-				android:layout_width="match_parent"
-				android:layout_height="1dp"
-				android:background="#50000000" />
-
 			<TextView
 				android:id="@+id/rtu_ui_v2_topbar"
 				android:layout_width="match_parent"
